refactor(Register): move form to separate component
